+++ examples/resources/biganimal_faraway_replica/import.sh
@@ -1,0 +1,2 @@
+# terraform import biganimal_faraway_replica.<resource_name> <project_id>/<cluster_id>
+terraform import biganimal_faraway_replica.faraway_replica prj_deadbeef01234567/p-abcd123456