+++ examples/resources/biganimal_pgd/aws/data_groups_with_witness_group/resource.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     biganimal = {
       source  = "EnterpriseDB/biganimal"
-      version = "0.8.1"
+      version = "1.0.0"
     }
     random = {
       source  = "hashicorp/random"
@@ -78,14 +78,14 @@ resource "biganimal_pgd" "pgd_cluster" {
         size              = "32 Gi"
       }
       pg_type = {
-        pg_type_id = "epas"
+        pg_type_id = "epas" #valid values ["epas", "pgextended", "postgres]"
       }
       pg_version = {
         pg_version_id = "15"
       }
       private_networking = false
       cloud_provider = {
-        cloud_provider_id = "aws"
+        cloud_provider_id = "bah:aws" // "bah:aws" uses BigAnimal's cloud account AWS, use "aws" for your cloud account
       }
       region = {
         region_id = "eu-west-1"
@@ -95,6 +95,9 @@ resource "biganimal_pgd" "pgd_cluster" {
         start_day  = 1
         start_time = "13:00"
       }
+      # pe_allowed_principal_ids = [
+      #   <example_value> # ex: 123456789012
+      # ]
     },
     {
       allowed_ip_ranges = [
@@ -132,14 +135,14 @@ resource "biganimal_pgd" "pgd_cluster" {
         size              = "32 Gi"
       }
       pg_type = {
-        pg_type_id = "epas"
+        pg_type_id = "epas" #valid values ["epas", "pgextended", "postgres]"
       }
       pg_version = {
         pg_version_id = "15"
       }
       private_networking = false
       cloud_provider = {
-        cloud_provider_id = "aws"
+        cloud_provider_id = "bah:aws" // "bah:aws" uses BigAnimal's cloud account AWS, use "aws" for your cloud account
       }
       region = {
         region_id = "eu-west-2"
@@ -149,6 +152,9 @@ resource "biganimal_pgd" "pgd_cluster" {
         start_day  = 2
         start_time = "15:00"
       }
+      # pe_allowed_principal_ids = [
+      #   <example_value> # ex: 123456789012
+      # ]
     }
   ]
   witness_groups = [
@@ -157,7 +163,7 @@ resource "biganimal_pgd" "pgd_cluster" {
         region_id = "us-east-1"
       }
       cloud_provider = {
-        cloud_provider_id = "aws"
+        cloud_provider_id = "bah:aws" // "bah:aws" uses BigAnimal's cloud account AWS, use "aws" for your cloud account
       }
       maintenance_window = {
         is_enabled = true